More attempts to parse the norm_obj.mat file
Still not successful... element ['s'] it appears to be a
zero-dimensional ndarray with all the data in it, but I can't seem to
drill down to the next level to get to that data other than to display
it in the shell.  I'll keep playing with it, I'm sure it's just a simple
thing.