Regression fixes: e.g. EIGENWORM_FILE
Fixed a couple of things that had broken the examples.
Sorry @JimHokanson, I am too allergic to typing literal file names in a
source file to not have a config.EIGENWORM_FILE.  I know that it's only
used in a couple of places but it's conceivable that more tools might
have to load the file, and so to me it's nice to have it in a constant
rather than hardcoded right in the file.  :)