rename BasicWorm
Basically I split the functions of "BasicWorm" into
NormalizedSkeletonAndContour, which is still a base class for
NormalizedWorm (and might as Jim suggests be removed), and
GeneralizedSkeletonAndContour, which can be loaded from an old Schafer
File.  The key difference between these two classes is that the latter
class makes no assumption about each frame having the same number of
points.