Working on statistics
For the first time, I was able to get a summary statistic comparing a
BasicWorm with a group of control feature files: add the computer vision
step and we have our Minimum Viable Product!

Run `video_to_features.py` to see this result:

`Comparison p and q values are 0.11 and 0.00, respectively.`

Still lots of bugs and the above values are certainly wrong.